feat: move Event to `domain.model` package
