+++ piikii-application/src/main/kotlin/com/piikii/application/domain/model/Event.kt
@@ -1,0 +1,17 @@
+package com.piikii.application.domain.model
+
+fun interface Event {
+    fun getTopic(): Event
+}
+
+class UserCreatedEvent : Event {
+    override fun getTopic(): Event {
+        TODO("Not yet implemented")
+    }
+}
+
+class UserUpdatedEvent : Event {
+    override fun getTopic(): Event {
+        TODO("Not yet implemented")
+    }
+}
