--- piikii-application/src/main/kotlin/com/piikii/application/port/output/eventbroker/UserCreatedEvent.kt
@@ -1,8 +1,0 @@
-package com.piikii.application.port.output.eventbroker
-
-class UserCreatedEvent : Event {
-
-    override fun getTopic(): Event {
-        TODO("Not yet implemented")
-    }
-}
